Refactor carousel to use swiper instead of react-swipeable-views
