+++ TODO.md
@@ -2,7 +2,6 @@
 
 - Remove material-ui. The following components are still used
   - Container
-- Replace outdated react-swipeable-views with something like swiper-react (npm i swiper)
 - Fix styles
 - Fix some unnecessarily complicated code
 - Generate CNAME automatically on deploy
