+++ package.json
@@ -17,7 +17,6 @@
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "@types/react-google-recaptcha": "^2.1.5",
-    "@types/react-swipeable-views": "^0.13.1",
     "@types/styled-components": "^5.1.25",
     "emailjs-com": "^2.6.4",
     "formik": "^2.2.9",
@@ -26,8 +25,8 @@
     "react-dom": "^18.2.0",
     "react-google-recaptcha": "^2.1.0",
     "react-scripts": "^5.0.1",
-    "react-swipeable-views": "^0.14.0",
     "styled-components": "^5.3.5",
+    "swiper": "^8.3.2",
     "typescript": "~4.7.4"
   },
   "scripts": {
