+++ src/components/index.ts
@@ -1,5 +1,4 @@
 export { default as PrimaryButton } from "./basic/PrimaryButton";
 export { default as Divider } from "./basic/Divider";
 export { default as FadeIn } from "./basic/FadeIn";
-export { default as Stepper } from "./basic/Stepper";
 export { default as Footer } from "./Footer";
