+++ src/components/sections/menu/Menu.tsx
@@ -32,6 +32,7 @@ const ButtonContainer = styled.div`
   position: fixed;
   top: 20px;
   right: 20px;
+  z-index: 1;
 `;
 
 interface ToggleMenuButtonProps {
